requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
Django==4.2.17
django-utils-six==2.0
durationpy==0.9
google-auth==2.37.0
gunicorn==23.0.0
idna==3.10
kubernetes==31.0.0
oauthlib==3.2.2
packaging==24.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.1
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
